Сделать меню частью PseudoPopup
